@@ -1,24 +1,47 @@
 {
-  "name": "frontend",
-  "version": "0.1.0",
+  "name": "@starport/template",
+  "version": "0.1.49",
+  "description": "A Vue 3 boilerplate project utilizing @starport/vue and @starport/vuex",
+  "author": "Tendermint, Inc <hello@tendermint.com>",
   "private": true,
   "scripts": {
     "dev": "npm install && vue-cli-service serve",
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "lint": "vue-cli-service lint",
+    "build": "vue-cli-service build",
+    "test": "echo \"Error: run tests from root\" && exit 1"
   },
   "dependencies": {
-    "@tendermint/vue": "0.1.12",
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "@cosmjs/launchpad": "0.25.0",
+    "@cosmjs/proto-signing": "0.25.0",
+    "@cosmjs/stargate": "0.25.0",
+    "@starport/vue": "^0.1.49",
+    "@starport/vuex": "^0.1.49",
+    "@vue/compiler-sfc": "^3.0.11",
+    "bip39": "^3.0.4",
+    "core-js": "^3.11.2",
+    "crypto-js": "^4.0.0",
+    "file-saver": "^2.0.5",
+    "vue": "^3.0.11",
+    "vue-router": "^4.0.3",
+    "vue-uuid": "^2.0.2",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.4.0",
-    "@vue/cli-plugin-router": "^4.4.0",
-    "@vue/cli-plugin-vuex": "^4.4.6",
-    "@vue/cli-service": "^4.4.0",
-    "vue-template-compiler": "^2.6.11"
+    "@babel/core": "^7.14.0",
+    "@babel/node": "^7.13.13",
+    "@vue/cli-plugin-babel": "^4.5.12",
+    "@vue/cli-plugin-eslint": "^4.5.12",
+    "@vue/cli-plugin-vuex": "^4.5.12",
+    "@vue/cli-service": "^4.5.12",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.8.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-vue": "^7.9.0",
+    "node-sass": "^4.12.0",
+    "prettier": "^2.2.1",
+    "protobufjs": "^6.11.2",
+    "sass-loader": "^8.0.2"
   }
 }
