@@ -1,12 +1,9 @@
-import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
-import store from "./store";
+import { createApp } from 'vue'
+import App from './App.vue'
+import store from './store'
+import router from './router'
+import vueLib from '@starport/vue'
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount("#app");
+const app = createApp(App)
+app.config.globalProperties._depsLoaded = true
+app.use(store).use(router).use(vueLib).mount('#app')
