@@ -1,21 +1,49 @@
 <template>
-  <div>
-    <router-view />
-  </div>
+	<div v-if="initialized">
+		<SpWallet ref="wallet" v-on:dropdown-opened="$refs.menu.closeDropdown()" />
+		<SpLayout>
+			<template v-slot:sidebar>
+				<Sidebar />
+			</template>
+			<template v-slot:content>
+				<router-view />
+			</template>
+		</SpLayout>
+	</div>
 </template>
 
 <style>
-.sp-container {
-  margin: 0 auto;
-  max-width: 800px;
-  padding: 1rem;
+body {
+	margin: 0;
 }
 </style>
 
 <script>
+import './scss/app.scss'
+import '@starport/vue/lib/starport-vue.css'
+import Sidebar from './components/Sidebar'
+
 export default {
-  created() {
-    this.$store.dispatch("cosmos/init");
-  },
-};
+	components: {
+		Sidebar
+	},
+	data() {
+		return {
+			initialized: false
+		}
+	},
+	computed: {
+		hasWallet() {
+			return this.$store.hasModule(['common', 'wallet'])
+		}
+	},
+	async created() {
+		await this.$store.dispatch('common/env/init')
+		this.initialized = true
+	},
+	errorCaptured(err) {
+		console.log(err)
+		return false
+	}
+}
 </script>
