@@ -1,5 +1,15 @@
+const path = require('path')
+
 module.exports = {
-  devServer: {
-    disableHostCheck: true,
-  },
-};
+	devServer: {
+		disableHostCheck: true
+	},
+	configureWebpack: {
+		resolve: {
+			symlinks: false,
+			alias: {
+				vue$: path.resolve('./node_modules/vue/dist/vue.esm-bundler.js')
+			}
+		}
+	}
+}
